move all instagram things to instagram app
