+++ src/branding/wagtail_hooks.py
@@ -1,8 +1,8 @@
 from wagtail.contrib.modeladmin.options import ModelAdmin, \
     modeladmin_register, ThumbnailMixin, ModelAdminGroup
 from django.utils.translation import ugettext_lazy as _
-
-from branding.models import Logo, InstagramFeed
+from instagram.models import InstagramFeedAdmin
+from branding.models import Logo
 
 
 class LogoAdmin(ThumbnailMixin, ModelAdmin):
@@ -12,13 +12,6 @@ class LogoAdmin(ThumbnailMixin, ModelAdmin):
     list_display = ('admin_thumb', 'category', 'link')
     thumb_image_field_name = 'logo'
     thumb_image_filter_spec = 'fill-128x128'
-
-
-class InstagramFeedAdmin(ModelAdmin):
-    model = InstagramFeed
-    menu_icon = 'fa-instagram'
-    menu_order = 110
-    create_template_name = "instagramfeed/create.html"
 
 
 class BrandingAdminGroup(ModelAdminGroup):
